Put back latching of play button (fixed on Mac) Fixed float/int issue with sliders and knobs (the decimal part of ranges was being discarded) Implemented value ioText type, which can receive values from outvalue opcode.

+++ src/quteknob.h
@@ -31,7 +31,7 @@ class QuteKnob : public QuteWidget
     ~QuteKnob();
 
     double getValue();
-    void setRange(int min, int max);
+    void setRange(double min, double max);
     void setValue(double value);
     virtual void setWidgetLine(QString line);
     QString getWidgetLine();
